partial checkin for implementing AdminEdit on CoreValues
