+++ src/components/Mantras/Mantras.jsx
@@ -80,7 +80,7 @@ function Mantras(props) {
                 <AdminEdits
                   page_id={page_id}
                   html_id={"header"}
-                  default_value={"Words To Live By"}
+                  default_value={`Words To Live By`}
                 />
               </h1>
               <h3>
